clippy/pyproject.toml
[project]
name = "clippy"
version = "0.1.0"
description = "macOS clipboard history with global hotkey"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
]

[project.scripts]
clippy = "clippy.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
